:basketball:

+++ lintcode/HouseRobber.java
@@ -4,8 +4,8 @@
 public class HouseRobber {
 
     public long houseRobber(int[] A) {
-        int prev = 0;
-        int cur = 0;
+        long prev = 0;
+        long cur = 0;
         for (int a : A) {
             int tmp = cur;
             cur = Math.max(prev + a, cur);
